Add files via upload
This is a Object detection computer vision application which is used to detect the particuller object in a web frame. this application is able to detect all kind of Objects of any colour. i have added a trackbar in this application so, we can make changes in the values of colours anytime according to the figure and hue of objects. we have used OpenCv and Numpy python libraries in this application.